feat(calls,chat,media): 1:1 audio/video qo'ng'iroq, dumaloq video xabar va rezyumlanadigan yuklash
Backend `calls/` handoff'i (`handoff/09-CALLS-*.md`) va 2026-08-03 dagi yangi OpenAPI
spec'i to'liq ulandi.

**Qo'ng'iroq** — yangi `:dev:feature:calls:{domain,data,presentation}`.

- `/calls` Socket.IO namespace'i (chatdan ALOHIDA: SDP hech qachon chat kanaliga
  tushmaydi), 17 hodisa, payload cheklari va ack shakli
- Holat mashinasi: RINGING → CONNECTING → ACTIVE, taymerlar 45/30/4h/20s, glare,
  band holat, ko'p qurilma (`call:taken`)
- Media qatlami — **Android'da haqiqiy WebRTC** (`io.github.webrtc-sdk:android`):
  `relayOnly` → `iceTransportPolicy = RELAY` (host/srflx umuman yig'ilmaydi), Opus
  FEC/DTX, old plan xizmati (usiz Android 14+ fonda mikrofonni jimgina o'chiradi).
  iOS'da `CallEngine` stub — `WebRTC.framework` alohida ish, lekin u jimgina
  yiqilmaydi: `call:invite` gacha yetib bormaydi va server chegaralari sarflanmaydi
- REST: `ice-servers` (kesh + **503 `NOT_IMPLEMENTED` alohida qayta ishlanadi**),
  `GET /v1/calls`, `POST /v1/calls/{id}/stats` — o'lchov **tanlangan juftlikdan**
  olinadi, javobsiz qo'ng'iroqda umuman yuborilmaydi
- UI: qo'ng'iroq ekrani ilova ustida (`CallHost`), chat sarlavhasida ikkita tugma,
  chat lentasida `CALL` pufakchasi, qo'ng'iroqlar tarixi

**`PREREQUISITES.md` ning uchala majburiy ishi bajarildi.** Eng muhimi — enum
chidamliligi: `cleanSwagger` ning 11-qadami kengayadigan enum'larni (`MessageTypeDto`,
`MediaKindDto`, qo'ng'iroq enum'lari…) Kotlin enum emas, `String` qilib generatsiya
qiladi. kotlinx noma'lum enum qiymatida BUTUN javobni yiqitadi; `coerceInputValues`
yaramaydi — u odatiy qiymat talab qiladi, generator esa uni qo'ymaydi.

⚠️ Allaqachon tarqatilgan klientda enum qat'iy, ya'ni backendning `CALL` yozadigan
deploy'i yangi versiya tarqalgandan KEYIN chiqishi kerak — `CALLS_PREREQUISITES_RESPONSE.md`.

**Media** (yangi spec bilan kelgan qismi):

- Hajm chegaralari olib tashlandi (yagona istisno `VIDEO_NOTE`/`STORY_VIDEO`)
- Rezyumlanadigan yuklash: `upload/init` → `part/{i}` → `complete`; ~10 MB dan
  kattasi avtomatik shu yo'ldan ketadi va uzilishdan keyin davom etadi
- **Dumaloq video xabar** — qabul qilish ham, yozib olish ham: kamera → dumaloq
  ko'rish varag'i → yuborish; kvadratga kesish, 60 s ga qirqish va 12 MB ga sig'adigan
  bitreyt yuborilgandan keyin, yuklash halqasi ichida (media3 `Presentation`)
- `quality` (AUTO/HIGH/ORIGINAL), `IMAGE_ORIGINAL`, `transcript`/`variants`,
  waveform 48 → 100 nuqta (son hech qayerda qotirilmaydi)

⚠️ `CAMERA` ruxsati manifestga qo'shilgani tizim kamerasini ochishni ham himoyalangan
qiladi — `rememberVideoCapture` endi ruxsatni tugma bosilganda so'raydi, aks holda
mavjud «Kamera» tugmasi `SecurityException` bilan yiqilardi.

`AppException` ga `errorCode` qo'shildi: backend konvertidagi `error.code` endi typed
xatogacha yetadi (403 «bog'lanmagan» mi yoki «bloklangan» mi, 503 «server yiqildi» mi
yoki «qo'ng'iroq o'chirilgan» mi — statusning o'zi buni aytmaydi).

`GET /v1/stories/archive` spec'da SAQLAB QOLINDI: uni biz qo'shgan edik, backend hali
chiqarmagan. DB: `28.sqm` (qo'ng'iroq ustunlari + `attachmentTranscript`), sxema 29.

+++ dev/core/uikit/src/androidMain/kotlin/dev/core/uikit/media/FilePicker.android.kt
@@ -57,7 +57,7 @@ private fun Context.readDocument(uri: Uri): PickedFile? {
     // yuklash ilovani o'ldiradi.
     if (meta.size != null && meta.size > MAX_FILE_BYTES) return null
 
-    val bytes = contentResolver.openInputStream(uri)?.use { it.readAtMost(MAX_FILE_BYTES) }
+    val bytes = contentResolver.openInputStream(uri)?.use { it.readAtMost(MAX_FILE_BYTES.toInt()) }
         ?: return null
     if (bytes.isEmpty()) return null
